inline resolver helper functions
